+++ NOrganizze/Transactions/TransactionCreateOptions.cs
@@ -29,10 +29,10 @@ namespace NOrganizze.Transactions
         public int? AmountCents { get; set; }
 
         [JsonProperty("category_id")]
-        public int? CategoryId { get; set; }
+        public long? CategoryId { get; set; }
 
         [JsonProperty("account_id")]
-        public int? AccountId { get; set; }
+        public long? AccountId { get; set; }
 
         [JsonProperty("paid")]
         public bool? Paid { get; set; }
