+++ NOrganizze/Transfers/Transfer.cs
@@ -10,7 +10,7 @@ namespace NOrganizze.Transfers
     public class Transfer : Transactions.Transaction
     {
         [JsonProperty("recurrence_id")]
-        public int? RecurrenceId { get; set; }
+        public long? RecurrenceId { get; set; }
 
         [JsonProperty("attachments")]
         public List<object> Attachments { get; set; }
