+++ NOrganizze/Transfers/TransferService.cs
@@ -22,12 +22,12 @@ namespace NOrganizze.Transfers
             return GetAsync<List<Transfer>>(Transfers, requestOptions, cancellationToken);
         }
 
-        public Transfer Get(int id, RequestOptions requestOptions = null)
+        public Transfer Get(long id, RequestOptions requestOptions = null)
         {
             return Get<Transfer>($"{Transfers}/{id}", requestOptions);
         }
 
-        public Task<Transfer> GetAsync(int id, RequestOptions requestOptions = null, CancellationToken cancellationToken = default)
+        public Task<Transfer> GetAsync(long id, RequestOptions requestOptions = null, CancellationToken cancellationToken = default)
         {
             return GetAsync<Transfer>($"{Transfers}/{id}", requestOptions, cancellationToken);
         }
@@ -42,22 +42,22 @@ namespace NOrganizze.Transfers
             return PostAsync<Transfer>(Transfers, options, requestOptions, cancellationToken);
         }
 
-        public Transfer Update(int id, TransferUpdateOptions options, RequestOptions requestOptions = null)
+        public Transfer Update(long id, TransferUpdateOptions options, RequestOptions requestOptions = null)
         {
             return Put<Transfer>($"{Transfers}/{id}", options, requestOptions);
         }
 
-        public Task<Transfer> UpdateAsync(int id, TransferUpdateOptions options, RequestOptions requestOptions = null, CancellationToken cancellationToken = default)
+        public Task<Transfer> UpdateAsync(long id, TransferUpdateOptions options, RequestOptions requestOptions = null, CancellationToken cancellationToken = default)
         {
             return PutAsync<Transfer>($"{Transfers}/{id}", options, requestOptions, cancellationToken);
         }
 
-        public Transfer Delete(int id, RequestOptions requestOptions = null)
+        public Transfer Delete(long id, RequestOptions requestOptions = null)
         {
             return Delete<Transfer>($"{Transfers}/{id}", requestOptions);
         }
 
-        public Task<Transfer> DeleteAsync(int id, RequestOptions requestOptions = null, CancellationToken cancellationToken = default)
+        public Task<Transfer> DeleteAsync(long id, RequestOptions requestOptions = null, CancellationToken cancellationToken = default)
         {
             return DeleteAsync<Transfer>($"{Transfers}/{id}", requestOptions, cancellationToken);
         }
